+++ src/PersonObjects/Sleeve/Sleeve.ts
@@ -534,6 +534,7 @@ export class Sleeve extends Person {
       }
       case SleeveTaskType.Recovery:
         this.shock = Math.min(100, this.shock + 0.0002 * cyclesUsed);
+        if (this.shock >= 100) this.resetTaskStatus();
         break;
       case SleeveTaskType.Synchro:
         this.sync = Math.min(100, this.sync + p.getIntelligenceBonus(0.5) * 0.0002 * cyclesUsed);