+++ src/index.tsx
@@ -3,7 +3,6 @@ import ReactDOM from "react-dom";
 
 import { TTheme as Theme, ThemeEvents, refreshTheme } from "./ui/React/Theme";
 import { LoadingScreen } from "./ui/LoadingScreen";
-import "./engineStyle";
 
 ReactDOM.render(
   <Theme>